// src/types/MicrobeConfig.ts
export interface MicrobeConfig {
  INITIAL_RADIUS: number;
  MAX_RADIUS: number;
  MIN_RADIUS: number;
  SPEED: number;
  MATING_DISTANCE: number;
  GESTATION_PERIOD: number;
  HUNGER_LIMIT: number;
  HUNGER_THRESHOLD_FOR_MATING: number;
  FULLNESS_DURATION: number;
  SIZE_INCREMENT: number;
  SIZE_DECREMENT: number;
  AGE_LIMIT: number;
  REST_CHANCE: number;
  REST_MIN_DURATION: number;
  REST_MAX_DURATION: number;
} 